added new scripts run_unet_1a.sh as an example to do training
